Moved VWProjectRisksTable to the main folder
